Refactor compression middleware to enhance gzip acceptance logic
- Updated the `compression` middleware to improve the logic for determining if gzip encoding is accepted by the client, using a new helper function `accepts_gzip_encoding`.
- Added a check to prevent compression if the response already has a `Content-Encoding` header, ensuring efficient response handling.
- Reintroduced tests for content type compression and compression configuration defaults in the middleware tests, enhancing test coverage.
